UI Barrinha

+++ jam161021/Assets/Scripts/BarScript.cs
@@ -7,11 +7,11 @@ public class BarScript : MonoBehaviour
 {
     public Slider slider;
    
-    public void setValue(int value){
+    public void setValue(float value){
         slider.value = value;
     }
-    
-    public void setMax(int value){
+
+    public void setMax(float value){
         slider.maxValue = value;
     }
 }
